+++ src/compas_viewer/config.py
@@ -251,7 +251,7 @@ class SidebarConfig(ConfigBase):
                 "type": "Sceneform",
                 "columns": [
                     {"title": "Name", "type": "label", "text": lambda obj: obj.name},
-                    {"title": "Show", "type": "checkbox", "action": lambda obj: obj.show, "checked": lambda obj, checked: setattr(obj, "show", checked)},
+                    {"title": "Show", "type": "checkbox", "checked": lambda obj: obj.show, "action": lambda obj, checked: setattr(obj, "show", checked)},
                 ],
             },
         ]